Split class I neural network helpers
