+++ mhcflurry/class1_neural_network.py
@@ -18,7 +18,6 @@ import gc
 import time
 import collections
 import json
-import multiprocessing
 import weakref
 import itertools
 import os
@@ -29,7 +28,7 @@ import pandas
 import torch
 import torch.nn as nn
 
-from . import amino_acid
+from . import pytorch_sizing
 from .hyperparameters import HyperparameterDefaults
 from .encodable_sequences import EncodableSequences, EncodingError
 from .allele_encoding import AlleleEncoding
@@ -44,6 +43,30 @@ from .random_negative_peptides import (
     supports_device_random_negative_encoding,
 )
 from .class1_affinity_training_data import AffinityDeviceTrainingData
+from .class1_encoding import (
+    _peptide_torch_encoding_name,
+    _peptide_torch_encoding_shape,
+    _peptide_torch_encoding_table,
+    _peptide_uses_torch_encoding,
+    peptide_sequences_to_network_input,
+)
+from .class1_training import (
+    _StreamingBatchIterableDataset,
+    _batched_validation_loss,
+    _carry_forward_validation_loss,
+    _early_stop_reached,
+    _make_streaming_batch_dataloader,
+    _materialize_repeated_peptide_batch,
+    _move_fit_batch_to_device,
+    _should_validate_epoch,
+    _sync_mean_loss,
+    _timing_enabled,
+    _timing_start,
+    _timing_stop,
+    _torch_from_numpy,
+    _update_min_validation_loss,
+    _validation_interval_from_hyperparameters,
+)
 from .pytorch_training import (
     configure_matmul_precision,
     effective_validation_batch_size,
@@ -54,710 +77,16 @@ from .pytorch_training import (
 )
 
 
-DEFAULT_PREDICT_BATCH_SIZE = "auto"
-_AUTO_BATCH_MAX_ROWS = 1_000_000  # cap past which kernel-launch savings flatten
-_AUTO_BATCH_MIN_ROWS = 1024       # floor: avoid pathologically tiny batches
-_AUTO_BATCH_CPU_FALLBACK = 32_768 # CPU: large batches thrash L3 — stay modest
-_AUTO_BATCH_FREE_FRACTION = 0.5   # half of free VRAM is the working-set budget
-_MPS_PSUTIL_WARNED = False        # one-shot warning if psutil is missing on MPS
-if os.environ.get("MHCFLURRY_DEFAULT_PREDICT_BATCH_SIZE"):
-    DEFAULT_PREDICT_BATCH_SIZE = int(os.environ["MHCFLURRY_DEFAULT_PREDICT_BATCH_SIZE"])
-    logging.info(
-        "Configured default predict batch size: %s" % DEFAULT_PREDICT_BATCH_SIZE
-    )
-
-
-def _estimate_peak_bytes_per_row(model):
-    """Worst-case peak activation bytes per sample during a forward.
-
-    Walks the model's configured layers and returns the maximum hidden-
-    layer width (in fp32 bytes) × 2 (one input + one output of the
-    current layer stay live under torch's eval-time no_grad reuse). A
-    4× multiplier covers framework overhead, cuDNN scratch buffers,
-    and Python-side tensor bookkeeping. Used by ``compute_prediction_batch_size``.
-    """
-    if model is None:
-        return 32 * 1024  # conservative 32 KB/row fallback
-    # Class1ProcessingModel is convolutional over the full
-    # flank+peptide sequence. Its peak activation is not represented by
-    # the affinity model's peptide_dense_layers/dense_layers attributes,
-    # so compute it from the Conv1d sequence width. This keeps prediction
-    # autosizing sensitive to future flank length / filter count changes
-    # instead of falling back to the generic 1024-width estimate.
-    try:
-        conv1 = model.conv1
-        seq_len = (
-            int(model.n_flank_length)
-            + int(model.peptide_max_length)
-            + int(model.c_flank_length)
-        )
-        widths = [
-            seq_len * int(conv1.in_channels),
-            seq_len * int(conv1.out_channels),
-        ]
-        for convs_name in ("n_flank_post_convs", "c_flank_post_convs"):
-            for layer in getattr(model, convs_name, []):
-                widths.append(seq_len * int(layer.out_channels))
-        if getattr(model, "flanking_averages", False):
-            widths.append(seq_len * int(conv1.out_channels))
-        peak = max(widths)
-        return int(peak * 4 * 2 * 4)  # fp32 × 2 buffers × 4x safety
-    except (AttributeError, TypeError, ValueError):
-        pass
-    # MergedClass1NeuralNetwork wraps N sub-networks and runs each one's
-    # forward independently in a list comprehension, then combines the
-    # outputs. All N sub-networks' peak intermediates are alive
-    # simultaneously, so the per-row peak is the SUM of per-sub-network
-    # peaks, not the max. Without this, the merged ensemble's auto-sized
-    # batch overshoots VRAM by Nx (~8x for the production 8-network
-    # release ensemble) and OOMs in calibrate's cartesian forward.
-    sub_networks = getattr(model, "networks", None)
-    if sub_networks is not None and not hasattr(model, "peptide_encoding_shape"):
-        try:
-            return int(sum(
-                _estimate_peak_bytes_per_row(net) for net in sub_networks
-            ))
-        except (AttributeError, TypeError) as exc:
-            logging.warning(
-                "Could not estimate peak per-row bytes for merged ensemble; "
-                "falling back to per-network walk: %s",
-                exc,
-            )
-    widths = []
-    try:
-        lc_out_len = int(model.peptide_encoding_shape[0])
-        lc_out_ch = int(model.peptide_encoding_shape[1])
-        for lc_layer in model.lc_layers:
-            lc_out_ch = getattr(lc_layer, "out_channels", lc_out_ch)
-            try:
-                lc_out_len = int(lc_layer.output_length)
-            except AttributeError:
-                pass
-        widths.append(lc_out_len * lc_out_ch)
-        for layer in model.peptide_dense_layers:
-            widths.append(int(layer.out_features))
-    except AttributeError:
-        widths.append(1024)
-    try:
-        allele_out = None
-        if getattr(model, "allele_embedding", None) is not None:
-            allele_out = int(model.allele_embedding.weight.shape[1])
-        for layer in getattr(model, "allele_dense_layers", []):
-            allele_out = layer.out_features
-        if allele_out is not None:
-            widths.append(allele_out)
-    except AttributeError:
-        pass
-    try:
-        for layer in model.dense_layers:
-            widths.append(int(layer.out_features))
-    except AttributeError:
-        widths.append(1024)
-    peak = max(widths) if widths else 1024
-    return int(peak * 4 * 2 * 4)  # fp32 × 2 buffers × 4x safety
-
-
-def _free_device_memory_bytes(device):
-    """Best-effort free-memory query. Returns a conservative value when
-    the device doesn't expose a direct free-memory API.
-
-    CUDA: ``torch.cuda.mem_get_info`` (exposed free + reserved tracking).
-    MPS: Apple's ``recommended_max_memory`` on unified memory,
-        minus whatever the MPS driver has already handed us. Cap by
-        ``psutil`` available RAM when present so other apps aren't
-        evicted. Falls back to 4 GB if neither API is reachable.
-    CPU / unknown: 2 GB conservative budget (the helper short-circuits
-        for CPU anyway, but keep a sensible value in case callers pass
-        a foreign device).
-    """
-    import torch
-    if device.type == "cuda":
-        try:
-            free, _ = torch.cuda.mem_get_info(device)
-            return int(free)
-        except Exception:
-            props = torch.cuda.get_device_properties(device)
-            reserved = torch.cuda.memory_reserved(device)
-            return max(int(props.total_memory) - int(reserved), 0)
-    if device.type == "mps":
-        # Apple Silicon: unified memory, so "free VRAM" is better
-        # estimated from the MPS driver's recommended ceiling minus
-        # whatever it's already allocated. Ceiling is typically
-        # ~75-80% of total system RAM on M-series chips.
-        try:
-            ceiling = int(torch.mps.recommended_max_memory())
-        except Exception:
-            ceiling = 4 * (1 << 30)
-        allocated = 0
-        try:
-            allocated = int(torch.mps.driver_allocated_memory())
-        except Exception:
-            pass
-        free = max(ceiling - allocated, 0)
-        # Don't evict other apps: also cap by the OS-reported free
-        # RAM when psutil is available. This gets us a realistic
-        # "what's safe to claim right now" rather than the MPS
-        # driver's peak permission.
-        try:
-            import psutil
-            free = min(free, int(psutil.virtual_memory().available))
-        except ImportError:
-            try:
-                from .workload_planning import system_memory_info_gb
-                available_gb = system_memory_info_gb().get("available_gb")
-                if available_gb is not None:
-                    free = min(free, int(available_gb * (1 << 30)))
-            except Exception:
-                # psutil isn't a hard dep. Log once per process so the
-                # skip is visible rather than silent — without this cap
-                # the MPS driver's "recommended max" can exceed what's
-                # actually safe to claim alongside other apps.
-                global _MPS_PSUTIL_WARNED
-                if not _MPS_PSUTIL_WARNED:
-                    logging.warning(
-                        "psutil not available and OS memory fallback failed; "
-                        "MPS free-memory estimate will use "
-                        "torch.mps.recommended_max_memory alone, which may "
-                        "overshoot actual available RAM."
-                    )
-                    _MPS_PSUTIL_WARNED = True
-        except Exception:
-            # Any other psutil failure (broken install, etc.) — skip
-            # the cap but don't fail the whole batch-size query.
-            pass
-        return free if free > 0 else 4 * (1 << 30)
-    return 2 * (1 << 30)
-
-
-def compute_prediction_batch_size(
-        device,
-        model=None,
-        num_workers_per_gpu=1,
-        free_memory_fraction=_AUTO_BATCH_FREE_FRACTION,
-        max_rows=_AUTO_BATCH_MAX_ROWS,
-        min_rows=_AUTO_BATCH_MIN_ROWS,
-        cpu_fallback=_AUTO_BATCH_CPU_FALLBACK):
-    """Auto-size a prediction batch for ``device`` and ``model``.
-
-    Divides free VRAM by the per-row peak activation estimate for the
-    model, scales by 1/``num_workers_per_gpu`` so co-resident workers
-    don't step on each other's budget, caps at ``max_rows`` (past which
-    kernel-launch savings flatten out), floors at ``min_rows``.
-
-    CPU: returns ``cpu_fallback`` — large batches on CPU thrash L3
-    cache and rarely help for the small networks mhcflurry trains.
-    """
-    if device.type == "cpu":
-        return cpu_fallback
-    peak_bytes = _estimate_peak_bytes_per_row(model)
-    free = _free_device_memory_bytes(device)
-    workers = max(int(num_workers_per_gpu), 1)
-    budget = int(free * float(free_memory_fraction) / workers)
-    budget = max(budget, peak_bytes * min_rows)
-    rows = budget // peak_bytes
-    return int(max(min_rows, min(rows, max_rows)))
-
-
-def _env_workers_per_gpu(default=1):
-    """Read the ``MHCFLURRY_MAX_WORKERS_PER_GPU`` env var.
-
-    The local parallelism pool sets this in each training worker so
-    auto-sized batching + training-memory checks can partition VRAM
-    across co-resident workers without the caller wiring it explicitly.
-    """
-    value = os.environ.get("MHCFLURRY_MAX_WORKERS_PER_GPU")
-    if value:
-        try:
-            return max(int(value), 1)
-        except ValueError:
-            pass
-    return default
-
-
-def resolve_prediction_batch_size(
-        value, device, model=None, num_workers_per_gpu=1):
-    """Resolve an explicit int or ``"auto"`` to a concrete batch size.
-
-    Accepts ``None`` as a synonym for ``"auto"``. Propagates an
-    explicit int through unchanged so callers can always pin the size
-    when they know better than the heuristic.
-    """
-    if value in (None, "auto"):
-        return compute_prediction_batch_size(
-            device,
-            model=model,
-            num_workers_per_gpu=num_workers_per_gpu,
-        )
-    return int(value)
-
-
-# Training-time memory multiplier: inference keeps only activations of
-# the current layer alive (input + output), training keeps the whole
-# forward-pass activation stack for backward plus gradients plus
-# optimizer state. RMSProp/Adam each store 1-2× weights in moving
-# averages on top. 4× the inference peak is a conservative floor that
-# leaves headroom for cuDNN workspace + Python-side torch overhead.
-_TRAINING_PEAK_MULTIPLIER = 4
-
-
-def check_training_batch_fits(
-        requested_batch_size,
-        device,
-        model,
-        num_workers_per_gpu=1,
-        free_memory_fraction=0.5,
-        min_batch=64,
-        logger=None):
-    """Verify that ``requested_batch_size`` will fit on ``device``.
-
-    Training peak memory = activations kept alive across the forward
-    pass (for backward), plus gradients, plus optimizer state. That's
-    roughly ``4 × _estimate_peak_bytes_per_row`` (inference peak).
-
-    Returns ``(effective_batch_size, shrunk: bool)``. When the
-    requested batch is too large for the available VRAM — partitioned
-    across co-resident workers — the batch is shrunk to the largest
-    power-of-two that fits (floored at ``min_batch``) and a loud
-    warning is emitted via ``logger`` / stderr explaining that the
-    training dynamics (BN running stats, gradient noise scale) now
-    differ from what the caller configured.
-
-    CPU short-circuits — no OOM risk there that a size-based heuristic
-    can catch. Returns ``(requested_batch_size, False)`` in that case.
-    """
-    import sys
-    if device.type == "cpu" or requested_batch_size <= min_batch:
-        return int(requested_batch_size), False
-    peak_bytes = _estimate_peak_bytes_per_row(model) * _TRAINING_PEAK_MULTIPLIER
-    free = _free_device_memory_bytes(device)
-    workers = max(int(num_workers_per_gpu), 1)
-    budget = int(free * float(free_memory_fraction) / workers)
-    max_rows = max(budget // peak_bytes, min_batch)
-    requested_batch_size = int(requested_batch_size)
-    if requested_batch_size <= max_rows:
-        return requested_batch_size, False
-    shrunk = 1
-    while shrunk * 2 <= max_rows:
-        shrunk *= 2
-    shrunk = max(shrunk, min_batch)
-    message = (
-        "!!! TRAINING BATCH WILL NOT FIT !!!  "
-        "Requested minibatch_size=%d on %s with %d worker(s)/GPU. "
-        "Estimated need ~%.1f GB of %.1f GB free VRAM (per-worker budget "
-        "~%.1f GB). Shrinking to %d.  "
-        "This CHANGES TRAINING DYNAMICS: batch-norm running stats, "
-        "gradient noise scale, and effective learning-rate schedule "
-        "all depend on batch size. Re-check convergence before "
-        "trusting the trained model. To pin an explicit size and "
-        "silence this guard, set a minibatch_size the caller knows "
-        "fits." % (
-            requested_batch_size, device, workers,
-            requested_batch_size * peak_bytes / 1e9,
-            free / 1e9,
-            budget / 1e9,
-            shrunk,
-        )
-    )
-    if logger is not None:
-        logger.warning(message)
-    else:
-        logging.warning(message)
-    # Also scream to stderr so it's loud in the job log regardless of
-    # which logger config the caller uses.
-    print("\n" + message + "\n", file=sys.stderr, flush=True)
-    return int(shrunk), True
-
+DEFAULT_PREDICT_BATCH_SIZE = pytorch_sizing.DEFAULT_PREDICT_BATCH_SIZE
+_env_workers_per_gpu = pytorch_sizing._env_workers_per_gpu
+check_training_batch_fits = pytorch_sizing.check_training_batch_fits
+compute_prediction_batch_size = pytorch_sizing.compute_prediction_batch_size
+resolve_prediction_batch_size = pytorch_sizing.resolve_prediction_batch_size
 
 KERAS_BATCH_NORM_EPSILON = 1e-3
 # Keras uses moving = moving * 0.99 + batch * 0.01. PyTorch's momentum is the
 # new-batch coefficient, so the equivalent value is 0.01.
 KERAS_BATCH_NORM_MOMENTUM = 0.01
-
-
-def _identity_collate(sample):
-    """Return a pre-batched sample unchanged."""
-    return sample
-
-
-def _materialize_repeated_peptide_batch(batch):
-    """Return ``batch`` with compact repeated-peptide input expanded.
-
-    Pretrain batches can carry each unique peptide encoding once plus a
-    ``peptide_repeat_count`` telling the training forward path to form the
-    peptide × allele batch on device. LSUV/data-dependent initialization still
-    expects ordinary row-aligned numpy arrays, so materialize that rare
-    one-batch path here.
-    """
-    repeat_count = batch.get("peptide_repeat_count")
-    if repeat_count is None:
-        return batch
-    result = dict(batch)
-    result["peptide"] = numpy.repeat(
-        result["peptide"],
-        int(repeat_count),
-        axis=0,
-    )
-    del result["peptide_repeat_count"]
-    return result
-
-
-class _StreamingBatchIterableDataset(torch.utils.data.IterableDataset):
-    """IterableDataset backing ``fit_streaming_batches``.
-
-    Supports two input formats:
-    - Legacy raw tuples: ``(alleles, peptides, affinities)``
-    - Pre-encoded tuples: ``(x_dict, y)``
-
-    The pre-encoded path is what allows worker-side prefetch in
-    pretraining: the dataset can be pickled into spawned workers without
-    serializing the full ``Class1NeuralNetwork`` instance.
-
-    Picklability contract
-    ---------------------
-    When ``num_workers>0``, PyTorch's DataLoader pickles the dataset via
-    ``ForkingPickler`` to ship it to each spawned worker. Two classes of
-    attributes would break that:
-
-    - Live generator objects (from ``generator=iter(...)``). Generators
-      are intrinsically unpicklable.
-    - Bound methods of ``Class1NeuralNetwork`` (``peptides_to_network_input``,
-      ``allele_encoding_to_network_input``). Pickling a bound method
-      pickles ``self``, i.e. the entire network.
-
-    We sidestep both by dropping those references whenever they're
-    redundant: the factory path never reads ``self.generator`` (always
-    fresh-instantiates per worker), and the pre-encoded path never reads
-    the ``*_to_input`` callbacks. The live-generator / raw-tuple path is
-    single-process by construction — the downgrade lives in
-    ``fit_streaming_batches`` — so zeroing these out doesn't remove any
-    functionality.
-    """
-
-    def __init__(
-        self,
-        *,
-        generator,
-        generator_factory=None,
-        source_batches_are_encoded=False,
-        allele_encoding_to_input=None,
-        peptides_to_network_input=None,
-    ):
-        # Factory takes precedence: every call to ``_iter_source`` will
-        # invoke the factory and ignore ``self.generator``, so don't
-        # retain the live (unpicklable) generator object.
-        self.generator = None if generator_factory is not None else generator
-        self.generator_factory = generator_factory
-        self.source_batches_are_encoded = source_batches_are_encoded
-        # Pre-encoded batches never consume these callbacks. Drop them
-        # to keep ``ForkingPickler`` from dragging the whole
-        # ``Class1NeuralNetwork`` instance into every spawned worker.
-        if source_batches_are_encoded:
-            self.allele_encoding_to_input = None
-            self.peptides_to_network_input = None
-        else:
-            self.allele_encoding_to_input = allele_encoding_to_input
-            self.peptides_to_network_input = peptides_to_network_input
-
-    def _iter_source(self):
-        worker_info = torch.utils.data.get_worker_info()
-        if self.generator_factory is not None:
-            worker_id = worker_info.id if worker_info is not None else 0
-            num_workers = (
-                worker_info.num_workers if worker_info is not None else 1
-            )
-            return self.generator_factory(
-                worker_id=worker_id,
-                num_workers=num_workers,
-            )
-        if worker_info is not None:
-            raise ValueError(
-                "fit_streaming_batches with DataLoader workers requires a "
-                "picklable generator_factory."
-            )
-        return self.generator
-
-    def _normalize_item(self, item):
-        if self.source_batches_are_encoded:
-            x_dict, y = item
-        else:
-            if self.peptides_to_network_input is None:
-                raise ValueError(
-                    "peptides_to_network_input is required for raw "
-                    "fit_streaming_batches batches."
-                )
-            alleles, peptides, affinities = item
-            allele_input = None
-            if self.allele_encoding_to_input is not None:
-                allele_input, _ = self.allele_encoding_to_input(alleles)
-            x_dict = {
-                "peptide": self.peptides_to_network_input(peptides),
-            }
-            if allele_input is not None:
-                x_dict["allele"] = allele_input
-            y = from_ic50(affinities)
-
-        result = dict(x_dict)
-        if not isinstance(y, numpy.ndarray):
-            y = numpy.asarray(y)
-        if y.dtype != numpy.float32:
-            y = y.astype(numpy.float32)
-        result["y"] = y
-        return result
-
-    def __iter__(self):
-        for item in self._iter_source():
-            yield self._normalize_item(item)
-
-
-def _batch_value_to_device(value, device, *, non_blocking, cast_float):
-    """Move a batch value (numpy array or tensor) to ``device``.
-
-    When ``cast_float=True`` and the source is float64, cast to float32 on
-    the CPU *before* transfer. MPS refuses to host float64 tensors, and
-    transferring fp64 just to narrow it on-device wastes bandwidth anyway.
-    int8 / float32 inputs skip the CPU-side cast so the pattern of shipping
-    compact int encodings and widening on the torch device stays intact.
-
-    When ``cast_float=False`` the value is shipped through in its source
-    dtype — used by the device-side fixed peptide encoding path where
-    peptide tensors are (N, L) int indices that get widened to (N, L, V)
-    fp32 via embedding lookup inside the network's forward pass.
-    """
-    if isinstance(value, numpy.ndarray):
-        value = _torch_from_numpy(value)
-    if cast_float and value.dtype == torch.float64:
-        value = value.float()
-    value = value.to(device, non_blocking=non_blocking)
-    if cast_float:
-        value = value.float()
-    return value
-
-
-def _torch_from_numpy(value):
-    if not value.flags.writeable:
-        value = value.copy()
-    return torch.from_numpy(value)
-
-
-_DEVICE_PEPTIDE_ENCODING_TRUE_VALUES = {
-    "1",
-    "true",
-    "yes",
-    "on",
-    "auto",
-    "cpu",
-    "device",
-    "torch",
-    "gpu",
-    "mps",
-}
-_DEVICE_PEPTIDE_ENCODING_FALSE_VALUES = {
-    "0",
-    "false",
-    "no",
-    "off",
-    "numpy",
-    "legacy",
-}
-
-_warned_legacy_peptide_vector_encoding = []
-
-
-def _warn_legacy_peptide_vector_encoding(value):
-    """Warn once that the legacy dense-vector peptide path is gone."""
-    if not _warned_legacy_peptide_vector_encoding:
-        _warned_legacy_peptide_vector_encoding.append(True)
-        logging.warning(
-            "peptide_amino_acid_encoding_torch=%r is deprecated and ignored: "
-            "peptides are always index-encoded ((N, L) int8) and embedded on "
-            "device via a frozen table. The legacy dense-vector peptide "
-            "encoding path has been removed.", value)
-
-
-def _peptide_torch_encoding_name(hyperparameters):
-    """Return the fixed amino-acid encoding for the peptide embedding table.
-
-    Peptides are always index-encoded ((N, L) int8) and embedded via a frozen
-    ``torch.nn.functional.embedding`` table (device-agnostic: CUDA/MPS/CPU). The
-    table comes from ``amino_acid.get_vector_encoding_df`` and is registered as
-    a non-persistent buffer, so it moves with ``.to(device)`` but never trains
-    or bloats the NPZ weight list.
-
-    The legacy ``peptide_amino_acid_encoding_torch=False`` dense-vector path has
-    been removed; a falsy value is accepted but ignored (with a one-time
-    deprecation warning) so older configs still load. A non-default value names
-    the encoding directly; an unknown name still raises.
-    """
-    mode = hyperparameters.get("peptide_amino_acid_encoding_torch", True)
-    if isinstance(mode, str):
-        mode_normalized = mode.strip().lower()
-        if mode_normalized in _DEVICE_PEPTIDE_ENCODING_FALSE_VALUES:
-            _warn_legacy_peptide_vector_encoding(mode)
-        else:
-            try:
-                amino_acid.get_vector_encoding_df(mode)
-                return mode
-            except KeyError:
-                pass
-            if mode_normalized not in _DEVICE_PEPTIDE_ENCODING_TRUE_VALUES:
-                raise ValueError(
-                    "Unsupported peptide_amino_acid_encoding_torch value %r. "
-                    "Expected bool, a true/false string, or one of %s."
-                    % (mode, sorted(amino_acid.ENCODING_DATA_FRAMES))
-                )
-    elif not mode:
-        _warn_legacy_peptide_vector_encoding(mode)
-
-    peptide_encoding = hyperparameters.get("peptide_encoding", {})
-    encoding_name = peptide_encoding.get("vector_encoding_name", "BLOSUM62")
-    try:
-        amino_acid.get_vector_encoding_df(encoding_name)
-    except KeyError:
-        raise ValueError(
-            "Peptide encoding requires a fixed vector encoding with a torch "
-            "lookup table; got %r. Available: %s"
-            % (encoding_name, sorted(amino_acid.ENCODING_DATA_FRAMES))
-        ) from None
-    return encoding_name
-
-
-def _peptide_uses_torch_encoding(hyperparameters):
-    """Deprecated: peptides are always index-encoded now (always True)."""
-    return True
-
-
-def _peptide_torch_encoding_table(encoding_name):
-    """Return a float32 lookup table indexed by ``AMINO_ACID_INDEX``."""
-    return amino_acid.vector_encoding_index_table(encoding_name)
-
-
-def _peptide_torch_encoding_shape(index_shape, encoding_name):
-    """Expand an ``(L,)`` categorical peptide shape to ``(L, V)``."""
-    return (
-        index_shape[0],
-        amino_acid.vector_encoding_length(encoding_name),
-    )
-
-
-def _categorical_kwargs_for_peptide_encoding(peptide_encoding):
-    """Extract kwargs shared by vector and categorical peptide encoders."""
-    return {
-        key: value
-        for key, value in peptide_encoding.items()
-        if key in (
-            "alignment_method",
-            "left_edge",
-            "right_edge",
-            "max_length",
-            "trim",
-            "allow_unsupported_amino_acids",
-        )
-    }
-
-
-def peptide_sequences_to_network_input(
-        peptides,
-        peptide_encoding,
-        peptide_amino_acid_encoding_torch=True):
-    """Encode peptide strings to the representation consumed by a network.
-
-    This is the central peptide-only string-to-array conversion. Peptides are
-    always returned as compact ``(N, L)`` int8 amino-acid indices; the
-    fixed-vector lookup then happens through the network's frozen torch
-    embedding table. (``peptide_amino_acid_encoding_torch`` is accepted for
-    config compatibility but the legacy dense-vector path is gone.)
-    """
-    if not peptide_amino_acid_encoding_torch:
-        _warn_legacy_peptide_vector_encoding(peptide_amino_acid_encoding_torch)
-    encoder = EncodableSequences.create(peptides)
-    return (
-        encoder.variable_length_to_fixed_length_categorical(
-            **_categorical_kwargs_for_peptide_encoding(peptide_encoding)
-        )
-        .astype("int8", copy=False)
-    )
-
-
-def _timing_enabled():
-    """Return True when fine-grained training timing is enabled."""
-    return os.environ.get("MHCFLURRY_ENABLE_TIMING", "0") == "1"
-
-
-def _timing_synchronize(device):
-    """Synchronize asynchronous device work for accurate wall timing."""
-    if device.type == "cuda":
-        torch.cuda.synchronize(device)
-    elif device.type == "mps" and hasattr(torch, "mps"):
-        torch.mps.synchronize()
-
-
-def _timing_start(device, enabled):
-    if enabled:
-        _timing_synchronize(device)
-    return time.perf_counter()
-
-
-def _timing_stop(start, device, enabled):
-    if enabled:
-        _timing_synchronize(device)
-    return time.perf_counter() - start
-
-
-def _move_fit_batch_to_device(batch, device, *, non_blocking):
-    """Move a training batch dict to ``device``.
-
-    When ``batch["peptide"]`` is a 2D integer array the torch-side fixed
-    peptide encoding path is active — keep it as int so the network's
-    embedding lookup can consume it. A 3D integer array is a compact
-    vector-encoded cache payload; still widen that to fp32 on device.
-    Float arrays flow through as before.
-    """
-    peptide_source = batch["peptide"]
-    peptide_repeat_count = batch.get("peptide_repeat_count")
-    if isinstance(peptide_source, numpy.ndarray):
-        _peptide_is_indices = (
-            peptide_source.ndim == 2
-            and numpy.issubdtype(peptide_source.dtype, numpy.integer)
-        )
-    else:
-        _peptide_is_indices = (
-            hasattr(peptide_source, "dim")
-            and peptide_source.dim() == 2
-            and peptide_source.dtype in (
-                torch.int8, torch.int16, torch.int32, torch.int64, torch.uint8
-            )
-        )
-    peptide_batch = _batch_value_to_device(
-        peptide_source,
-        device,
-        non_blocking=non_blocking,
-        cast_float=not _peptide_is_indices,
-    )
-    y_batch = _batch_value_to_device(
-        batch["y"],
-        device,
-        non_blocking=non_blocking,
-        cast_float=False,
-    )
-    inputs = {"peptide": peptide_batch}
-    if peptide_repeat_count is not None:
-        inputs["peptide_repeat_count"] = int(peptide_repeat_count)
-    if "allele" in batch:
-        inputs["allele"] = _batch_value_to_device(
-            batch["allele"],
-            device,
-            non_blocking=non_blocking,
-            cast_float=True,
-        )
-    weights_batch = None
-    if "weight" in batch:
-        weights_batch = _batch_value_to_device(
-            batch["weight"],
-            device,
-            non_blocking=non_blocking,
-            cast_float=True,
-        )
-    return (inputs, y_batch, weights_batch)
 
 
 def _run_training_batch(
@@ -837,311 +166,6 @@ def _run_prepared_training_batches(
         "train_rows": train_rows,
         "first_batch_time": first_batch_time,
     }
-
-
-def _sync_mean_loss(losses, *, device, timing_enabled):
-    """Return mean detached loss with one device sync plus sync timing."""
-    loss_sync_start = _timing_start(device, timing_enabled)
-    train_loss = (
-        torch.stack(losses).mean().item()
-        if losses else float("nan")
-    )
-    loss_sync_time = _timing_stop(loss_sync_start, device, timing_enabled)
-    return train_loss, loss_sync_time
-
-
-def _validation_interval_from_hyperparameters(hyperparameters):
-    """Return normalized validation interval for both training paths."""
-    return max(1, int(hyperparameters.get("validation_interval", 1) or 1))
-
-
-def _early_stop_reached(
-    *,
-    epoch_index,
-    min_val_loss_epoch,
-    patience,
-    min_epochs=0,
-    early_stopping=True,
-    strict=True,
-):
-    """Shared early-stopping decision using zero-based epoch indices.
-
-    With ``strict=True`` (the affinity ``fit()`` path) training stops the
-    first epoch *strictly after* ``max(min_val_loss_epoch + patience,
-    min_epochs - 1)`` — i.e. ``patience`` no-improvement epochs followed by
-    one more before stopping. This matches the pre-2.3.0 ``fit()`` condition
-    ``epoch > min_val_loss_iteration + patience``.
-
-    With ``strict=False`` (the streaming pretrain path) training stops *at*
-    that threshold, reproducing the pre-2.3.0 streaming condition
-    ``epoch >= min_val_loss_iteration + patience`` (which used one-based
-    epochs). Keeping the streaming path inclusive means pretrained base
-    weights reproduce historical runs rather than training one extra epoch.
-    """
-    if not early_stopping or min_val_loss_epoch is None:
-        return False
-    threshold = max(
-        int(min_val_loss_epoch) + int(patience),
-        int(min_epochs) - 1,
-    )
-    if strict:
-        return int(epoch_index) > threshold
-    return int(epoch_index) >= threshold
-
-
-def _should_validate_epoch(
-    *,
-    validation_enabled,
-    epoch_index,
-    max_epochs,
-    validation_interval,
-    early_stopping,
-    min_val_loss_epoch,
-    patience,
-    min_epochs=0,
-    strict=True,
-):
-    """Shared validation cadence for fit() and fit_streaming_batches().
-
-    ``strict`` is forwarded to :func:`_early_stop_reached` so the "validate on
-    the epoch we're about to stop" rule uses the same inclusive/exclusive
-    threshold as the caller's break condition.
-    """
-    if not validation_enabled:
-        return False
-    is_last_epoch = int(epoch_index) == int(max_epochs) - 1
-    patience_would_trigger = _early_stop_reached(
-        epoch_index=epoch_index,
-        min_val_loss_epoch=min_val_loss_epoch,
-        patience=patience,
-        min_epochs=min_epochs,
-        early_stopping=early_stopping,
-        strict=strict,
-    )
-    return (
-        int(epoch_index) % int(validation_interval) == 0
-        or is_last_epoch
-        or patience_would_trigger
-    )
-
-
-def _carry_forward_validation_loss(fit_info):
-    """Keep val_loss aligned with epochs when validation is skipped."""
-    return fit_info["val_loss"][-1] if fit_info["val_loss"] else float("nan")
-
-
-def _update_min_validation_loss(
-    *,
-    val_loss,
-    epoch_index,
-    min_val_loss,
-    min_val_loss_epoch,
-    min_delta,
-):
-    """Update min validation state after a measured validation epoch."""
-    if min_val_loss is None or val_loss < min_val_loss - min_delta:
-        return val_loss, epoch_index
-    return min_val_loss, min_val_loss_epoch
-
-
-def _effective_num_workers(num_workers):
-    """Downgrade ``num_workers`` to 0 when running inside a daemon process.
-
-    Safety net for DataLoader's ``num_workers>0`` path, which requires
-    spawning child processes — forbidden from daemon parents. mhcflurry's
-    training orchestrator switched to ``NonDaemonPool`` (see
-    ``mhcflurry.parallelism``) so in production Pool workers are now
-    non-daemonic and this downgrade does NOT fire. It remains as
-    defense-in-depth for:
-
-      - External callers that wrap ``Class1NeuralNetwork.fit`` in a
-        daemonic process of their own (e.g. ``ProcessPoolExecutor`` is
-        non-daemon by default but ``multiprocessing.Pool`` without the
-        NonDaemonPool subclass is daemonic).
-      - Test suites that explicitly construct daemon subprocesses (our
-        own ``test_dataloader_num_workers_downgrades_in_daemon_context``
-        exercises this path).
-
-    Emits a ``DeprecationWarning`` — it's worth knowing if this fires,
-    since it means the caller is running in a daemon context and
-    DataLoader prefetch is silently disabled. Use ``warnings.filterwarnings``
-    to silence if intentional.
-    """
-    if num_workers > 0 and multiprocessing.current_process().daemon:
-        import warnings
-        warnings.warn(
-            f"dataloader_num_workers={num_workers} requested from a daemon "
-            f"process; downgrading to 0 because daemon processes cannot "
-            f"spawn children. If this is mhcflurry's Pool worker, switch "
-            f"to NonDaemonPool (mhcflurry/parallelism); if it's "
-            f"your own caller, use a non-daemonic executor or accept the "
-            f"downgrade (set dataloader_num_workers=0 explicitly to silence).",
-            DeprecationWarning,
-            stacklevel=3,
-        )
-        return 0
-    return num_workers
-
-
-def _make_streaming_batch_dataloader(
-    dataset,
-    num_workers,
-):
-    """Construct a DataLoader for ``fit_streaming_batches``.
-
-    The dataset already yields complete batches, so automatic batching is
-    disabled (``batch_size=None``) and we keep the payload as numpy
-    arrays via ``_identity_collate``. That avoids worker-side tensor
-    materialization and keeps the worker path picklable for spawned
-    subprocesses.
-    """
-    effective_workers = _effective_num_workers(num_workers)
-    kwargs = dict(
-        dataset=dataset,
-        batch_size=None,
-        num_workers=effective_workers,
-        collate_fn=_identity_collate,
-    )
-    if effective_workers > 0:
-        kwargs["multiprocessing_context"] = "spawn"
-        kwargs["prefetch_factor"] = 2
-    return torch.utils.data.DataLoader(**kwargs)
-
-
-_VALIDATION_INEQUALITY_FALLBACK_WARNED = False
-
-
-def _warn_validation_inequality_fallback_once(loss_name):
-    """Log once per process when val_loss drops to the single-shot inequality path.
-
-    The fallback exists to preserve 2.1.x semantics: ``MSEWithInequalities`` and
-    its multi-output variant exclude encoded ``2.0`` targets from the
-    denominator, and batchwise averaging would change that count. Users should
-    know that validation is then unbatched so they can size ``batch_size``
-    knowing the cost.
-    """
-    global _VALIDATION_INEQUALITY_FALLBACK_WARNED
-    if _VALIDATION_INEQUALITY_FALLBACK_WARNED:
-        return
-    _VALIDATION_INEQUALITY_FALLBACK_WARNED = True
-    logging.warning(
-        "validation loss running single-shot (no batching) because "
-        "%s has inequality targets that would change the legacy denominator "
-        "if split across batches. This log fires once per process.",
-        loss_name,
-    )
-
-
-def _batched_validation_loss(
-    *,
-    network,
-    eager_network,
-    val_peptide,
-    val_allele,
-    val_y,
-    val_weights,
-    loss_obj,
-    batch_size,
-):
-    """Compute validation loss with a fixed-size fast path and eager tail path.
-
-    Every ``network(inputs)`` call inside this helper sees the same
-    shape (``batch_size``), so GPU validation stays memory-bounded and
-    optional compiled-validation paths can specialize once and reuse.
-    That matters for H100 and A100 training where single-shot forward
-    passes over the full validation set (up to ~185K rows for pan-allele
-    training) were eating 15+ GB of VRAM per worker.
-
-    Semantics: iterate ``val_peptide`` in ``batch_size``-row chunks and
-    return the exact mean validation loss across all rows. Full-size
-    batches run through ``network`` (typically compiled); any short tail
-    batch runs through ``eager_network`` to avoid recompiling on an odd
-    shape.
-
-    Weighted validation and cross-example losses fall back to the
-    original single-shot computation to preserve exact semantics:
-    averaging per-batch weighted means is not the same as taking the
-    weighted mean over the whole validation set, and losses like
-    ``MultiallelicMassSpecLoss`` couple samples within a batch.
-
-    Returns a Python float (already detached from autograd).
-    """
-    n_val = val_peptide.shape[0]
-    if n_val == 0:
-        return float("nan")
-
-    # Short-circuit cheap predicates first so we only sync the device for the
-    # inequality-denominator check when none of the basic conditions apply.
-    needs_basic_fallback = (
-        n_val < batch_size
-        or val_weights is not None
-        or not getattr(loss_obj, "supports_independent_samples", True)
-    )
-    inequality_fallback = (
-        not needs_basic_fallback
-        and _validation_loss_has_legacy_inequality_denominator(loss_obj, val_y)
-    )
-    if needs_basic_fallback or inequality_fallback:
-        # Fallback for tiny val sets (typical in unit tests) and for
-        # weighted / cross-example / inequality losses whose reductions are
-        # not decomposable batchwise without changing the legacy denominator.
-        if inequality_fallback:
-            _warn_validation_inequality_fallback_once(type(loss_obj).__name__)
-        inputs = {"peptide": val_peptide}
-        if val_allele is not None:
-            inputs["allele"] = val_allele
-        preds = eager_network(inputs)
-        batch_loss = loss_obj(preds, val_y, sample_weights=val_weights)
-        return batch_loss.item()
-
-    n_full = (n_val // batch_size) * batch_size
-    loss_accum = 0.0
-    weight_accum = 0.0
-    for start in range(0, n_full, batch_size):
-        end = start + batch_size
-        inputs = {"peptide": val_peptide[start:end]}
-        if val_allele is not None:
-            inputs["allele"] = val_allele[start:end]
-        preds = network(inputs)
-        weights_slice = (
-            val_weights[start:end] if val_weights is not None else None
-        )
-        batch_loss = loss_obj(
-            preds, val_y[start:end], sample_weights=weights_slice
-        )
-        loss_accum += batch_loss.item() * batch_size
-        weight_accum += batch_size
-    if n_full < n_val:
-        inputs = {"peptide": val_peptide[n_full:n_val]}
-        if val_allele is not None:
-            inputs["allele"] = val_allele[n_full:n_val]
-        tail_loss = loss_obj(
-            eager_network(inputs),
-            val_y[n_full:n_val],
-            sample_weights=None,
-        )
-        tail_size = n_val - n_full
-        loss_accum += tail_loss.item() * tail_size
-        weight_accum += tail_size
-    return loss_accum / weight_accum
-
-
-def _validation_loss_has_legacy_inequality_denominator(loss_obj, val_y):
-    """Return True when batch-size weighting would change inequality loss.
-
-    ``MSEWithInequalities`` and its multi-output variant keep 2.1.x behavior:
-    encoded ``2.0`` targets are excluded from the denominator. If such rows
-    are split unevenly across validation batches, averaging batch means by raw
-    batch size changes ``val_loss``. Use the old single-shot path only for that
-    case; all other independent-sample inequality rows remain batchable.
-    """
-    if not getattr(loss_obj, "supports_inequalities", False):
-        return False
-    y_t = val_y.reshape(-1)
-    if getattr(loss_obj, "supports_multiple_outputs", False):
-        output_indices = (y_t / 10.0).long()
-        y_t = y_t - output_indices.to(y_t.dtype) * 10.0
-    return bool(torch.any(y_t == 2.0).item())
 
 
 class Class1NeuralNetworkModel(nn.Module):
